titanic-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.2
python-dotenv==1.0.1
pydantic==2.6.1
joblib==1.3.2

